pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "bsb-yaml"
authors = [{name = "Robin De Schepper", email = "robingilbert.deschepper@unipv.it"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"]
dynamic = ["version", "description"]
dependencies = ["bsb-core>=4.0.0b4,<=4.0.0b9999", "PyYAML~=6.0"]

[project.entry-points."bsb.config.parsers"]
yaml = "bsb_yaml.parser"

[project.entry-points."bsb.config.templates"]
yaml_templates = "bsb_yaml.templates"

[tool.flit.module]
name = "bsb_yaml"

[project.optional-dependencies]
test = ["bsb-core[parallel]", "bsb-test>=0.0.0b7,<=0.0.0b9999", "coverage~=7.0"]
dev = [
    "pre-commit~=3.5",
    "black~=23.11",
    "isort~=5.12",
    "bump-my-version~=0.18"
]

[tool.isort]
profile = "black"

[tool.bumpversion]
current_version = "0.0.0-b4"
parse = """(?x)
    (?P<major>0|[1-9]\\d*)\\.
    (?P<minor>0|[1-9]\\d*)\\.
    (?P<patch>0|[1-9]\\d*)
    (?:
        -                             # dash seperator for pre-release section
        (?P<pre_l>[a-zA-Z-]+)         # pre-release label
        (?P<pre_n>0|[1-9]\\d*)        # pre-release version number
    )?                                # pre-release section is optional
"""
serialize = [
    "{major}.{minor}.{patch}-{pre_l}{pre_n}",
    "{major}.{minor}.{patch}",
]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
tag = true
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
commit_args = "--no-verify"

[tool.bumpversion.parts.pre_l]
values = ["dev", "a", "b", "rc", "final"]
optional_value = "final"

[[tool.bumpversion.files]]
filename = "bsb_yaml/__init__.py"